Replace Alpaca with IG Markets broker integration for CFD trading
Rewrites the broker connector to use trading_ig library for IG Markets
live/demo accounts. Changes quantity types from int to float throughout
the pipeline to support fractional CFD sizes. Simplifies the WebSocket
feeder to yfinance-only polling (no broker credentials needed). Adds
EPIC resolution with 3-tier cache (in-memory, DynamoDB, IG API), and
updates CDK stacks to use IG credential templates.

+++ src/claudestreet/models/trade.py
@@ -32,7 +32,7 @@ class Order(BaseModel):
     id: str = Field(default_factory=lambda: uuid.uuid4().hex[:12])
     symbol: str
     side: OrderSide
-    quantity: int
+    quantity: float
     limit_price: float | None = None
     stop_price: float | None = None
     status: OrderStatus = OrderStatus.PENDING
@@ -46,7 +46,7 @@ class Position(BaseModel):
     id: str = Field(default_factory=lambda: uuid.uuid4().hex[:12])
     symbol: str
     side: OrderSide
-    quantity: int
+    quantity: float
     entry_price: float
     current_price: float = 0.0
     stop_loss: float = 0.0
